All changes are to make splitByColon work as expected.
Wrote step-by-stemp comments for Marshall's new char-by-char splitByDelim() method (which is quite nice).

Added an initial $inQuotes value in splitByDelim().

Removed $convertSingle paramter from splitByDelim() (we have agreed that all results are to be arrays, not strings).

In its place, added $recurse parameter to splitByDelim(), this tells the method whether or not to recursively parse past the first occurrence of the delimiter.

Added $recurse params and default vals to splitBySemi() (true), splitByComma (true), and splitByColon (false).

Tested it with sample vCard parse, appears to work as expected.